Lazy load the dashboard while the user logs in
